+++ src/rules/function-max-empty-lines/__test__/index.js
@@ -1,4 +1,4 @@
-import testRule from "../../../testUtils/stylelint-test-rule-tape"
+import testRule from "../../../testUtils/testRule"
 import rule, { ruleName, messages } from ".."
 
 testRule(rule, {