+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Head
 
 - Added: exposed `stylelint.createRuleTester()`.
+- Added: `function-max-empty-lines` rule.
 - Fixed: `property-value-blacklist` and `-whitelist` no longer error on properties without a corresponding list entry.
 - Fixed: `font-weight-notation` now ignores `initial` value.
 
@@ -15,7 +16,6 @@
 - Added: `no-indistinguishable-colors` rule.
 - Added: `at-rule-semicolon-newline-after` rule.
 - Added: `stylelint-disable-reason` rule.
-- Added: `function-max-empty-lines` rule.
 - Fixed: `no-duplicate-selectors` contextualizes selectors by all at-rules, not just media queries.
 - Fixed: `declaration-bang-space-*` understands arbitrary bang declarations (e.g. `!default`).
 - Fixed: `font-weight-notation` now ignore `inherit` value.